split benchmark and UT
