+++ bestla/CMakeLists.txt
@@ -64,7 +64,8 @@ endif()
 
 function(add_ut_flag UT_OPTION)
 	if(${${UT_OPTION}})
-	  target_compile_definitions(${PROJECT_NAME}_ut PRIVATE ${UT_OPTION})
+	  # target_compile_definitions(${PROJECT_NAME}_ut PRIVATE ${UT_OPTION})
+    add_compile_definitions(${UT_OPTION})
 	endif()
 endfunction()
 
@@ -96,8 +97,17 @@ if(UT_BUILD)
 	add_ut_flag(BTLA_UT_KERNEL_INTRIN)
 	add_ut_flag(BTLA_UT_KERNEL_JIT)
 	add_ut_flag(BTLA_UT_KERNEL_WRAPPER)
-	add_ut_flag(BTLA_UT_BENCHMARK)
-
 	target_link_libraries(${PROJECT_NAME}_ut PRIVATE ${PROJECT_NAME})
 endif(UT_BUILD)
 
+if(BTLA_UT_BENCHMARK)
+  file(GLOB srcs ${PROJECT_NAME}/ut/bestla_benchmark.cpp) #compile everything even run parts of UTs
+	file(GLOB ut_headers ${PROJECT_NAME}/ut/*.h)
+  include_directories(${PROJECT_NAME})
+	add_executable(${PROJECT_NAME}_benchmark ${srcs} ${headers} ${ut_headers})
+  if(BTLA_UT_OPENMP)
+    include(FindOpenMP)
+    target_compile_definitions(${PROJECT_NAME} INTERFACE BTLA_USE_OPENMP)
+    target_link_libraries(${PROJECT_NAME}_benchmark PRIVATE OpenMP::OpenMP_CXX)
+  endif()
+endif(BTLA_UT_BENCHMARK)
