+++ bestla/bestla/ut/bestla_wrapper.cpp
@@ -7,12 +7,6 @@ class UT_Fp32Fp32 {
  public:
   UT_Fp32Fp32() {
     UT_START();
-#ifdef BTLA_UT_BENCHMARK
-    benchmark_all(1, 4096, 4096, 32);
-    benchmark_all(1024, 4096, 4096, 32);
-    benchmark_all(2048, 4096, 4096, 32);
-#endif  // BTLA_UT_BENCHMARK
-
     CheckISA(AVX2);
     ut<sAVX2>(1, 1, 1);
     ut<sAVX2>(8, 48, 2);
@@ -53,75 +47,6 @@ class UT_Fp32Fp32 {
     parallel::GemmRun<Parallel>(launcher, args, UT_Threading::get());
     ut::buffer_error(ref.data(), matC.data(), ref.size(), 0.001f);
   }
-
-  using AType = float;
-  using BType = float;
-  using CType = float;
-  template <typename Core_T, typename LOG_T>
-  void benchmark(int m, int n, int k, int batch, AType* A, BType* B, CType* C, float timems, int threads) {
-    LOG_T log;
-    using Parallel = parallel::gemm::SchedulerBase<Core_T>;
-    using Launcher =
-        wrapper::gemm::LauncherBase<Core_T::ISA, Core_T, prologue_a::gemm::ActivationBase, prologue_b::gemm::WeightPack,
-                                    epilogue::gemm::AccumulatorWriteBackFp32>;
-    Launcher kernel;
-    UT_Threading::set_threads(threads);
-    auto corestr = gemm::CoreAttr::to_str(Core_T::ID);
-    utils::timer<std::chrono::milliseconds> tm;
-    auto tmpB = kernel.mProB.createStorage(n, k);
-    std::vector<storage::gemm::StoragePackedWeight> packBs(batch, 0);
-    std::vector<int8_t> bufB(tmpB.mSize * batch);
-    for (size_t i = 0; i < batch; i++) {
-      packBs[i] = tmpB;
-      packBs[i].assign(bufB.data() + i * tmpB.mSize);
-      kernel.mProB.packWeight(n, k, {B + i * n * k, n, &packBs[i]}, UT_Threading::get());
-    }
-    auto psize = (size_t)m * n * k * 2;
-    tm.start();
-    while (tm.stop() < timems) {
-      for (size_t i = 0; i < batch; i++) {
-        log.start();
-        utils::GemmProblem gp(1, m, n, k);
-        typename Launcher::Param args{gp, {A + i * m * k, k}, {0, 0, &packBs[i]}, {C + i * m * n, n}};
-        parallel::GemmRun<Parallel>(kernel, args, UT_Threading::get());
-        if (log.stop()) {
-          double flops = double(psize) / log.avg_val / 1e6;
-          printf("Threads %d %s %s Flops:%.3f PerCoreFlops:%.3f\n ", threads, corestr, log.get_log_str(), flops,
-                 flops / threads);
-        }
-      }
-    }
-  }
-
-  void benchmark_all(size_t m, size_t n, size_t k, size_t batch) {
-    printf("%s %d %d %d %d\n", __FUNCTION__, int(m), int(n), int(k), int(batch));
-    avector<AType> A(m * k * batch);
-    avector<BType> B(k * n * batch);
-    avector<CType> C(m * n * batch, 0), RefC(m * n * batch, 0);
-    fill_buffer_randn(A.data(), m * k, -0.5f, 0.5f);
-    fill_buffer_randn(B.data(), n * k, -0.5f, 0.5f);
-    for (size_t i = 0; i < batch - 1; i++) {
-      memcpy(A.data() + i * m * k, A.data(), m * k * sizeof(AType));
-      memcpy(B.data() + i * n * k, B.data(), n * k * sizeof(BType));
-    }
-    using LOG = timer_statistics_logger<100>;
-
-    float testtime = 500.f;
-    GetCPUDevice();
-    if (_cd->AVX512F()) {
-      benchmark<sAVX512F, LOG>(m, n, k, batch, A.data(), B.data(), C.data(), testtime, _cd->getThreads());
-      benchmark<sAVX512F, LOG>(m, n, k, batch, A.data(), B.data(), C.data(), testtime, _cd->getCores());
-    }
-    if (_cd->AVX2()) {
-      if (_cd->isHybrid()) {
-        benchmark<sAVX2, LOG>(m, n, k, batch, A.data(), B.data(), C.data(), testtime, _cd->getThreads());
-        benchmark<sAVX2, LOG>(m, n, k, batch, A.data(), B.data(), C.data(), testtime, _cd->getCores());
-        benchmark<sAVX2, LOG>(m, n, k, batch, A.data(), B.data(), C.data(), testtime, _cd->getPcoreNum());
-      } else {
-        benchmark<sAVX2, LOG>(m, n, k, batch, A.data(), B.data(), C.data(), testtime, _cd->getThreads());
-      }
-    }
-  }
 };
 #ifdef BTLA_UT_WRAPPER
 static UT_Fp32Fp32 sUT_Fp32Fp32;
@@ -132,10 +57,6 @@ class UT_U8S8S32 {
   UT_U8S8S32() {
     UT_START();
     GetCPUDevice();
-#ifdef BTLA_UT_BENCHMARK
-    benchmark_all(1024, 4096, 4096, 32);
-    benchmark_all(2048, 4096, 4096, 32);
-#endif
     if (_cd->AVX512_VNNI()) {
       ut<sAVX512_VNNI>(4, 48, 4);
       ut<sAVX512_VNNI>(1, 1, 1);
@@ -207,95 +128,6 @@ class UT_U8S8S32 {
     parallel::GemmRun<Parallel>(launcher, args, UT_Threading::get());
     ut::buffer_error(refC.data(), matC.data(), refC.size(), 0.001f);
   }
-
-  using AType = uint8_t;
-  using BType = int8_t;
-  using CType = int;
-  template <typename Core_T, typename LOG_T>
-  void benchmark(int m, int n, int k, int batch, AType* A, BType* B, CType* C, float timems, int threads) {
-    LOG_T log;
-    using Parallel = parallel::gemm::SchedulerBase<Core_T>;
-    using Launcher =
-        wrapper::gemm::LauncherBase<Core_T::ISA, Core_T, prologue_a::gemm::ActivationBase, prologue_b::gemm::WeightPack,
-                                    epilogue::gemm::AccumulatorWriteBackInt32>;
-    Launcher kernel;
-    UT_Threading::set_threads(threads);
-    auto corestr = gemm::CoreAttr::to_str(Core_T::ID);
-    utils::timer<std::chrono::milliseconds> tm;
-    auto tmpB = kernel.mProB.createStorage(n, k);
-    std::vector<storage::gemm::StoragePackedWeight> packBs(batch, 0);
-    std::vector<int8_t> bufB(tmpB.mSize * batch);
-    for (size_t i = 0; i < batch; i++) {
-      packBs[i] = tmpB;
-      packBs[i].assign(bufB.data() + i * tmpB.mSize);
-      kernel.mProB.packWeight(n, k, {B + i * n * k, n, &packBs[i]}, UT_Threading::get());
-    }
-    auto psize = (size_t)m * n * k * 2;
-    tm.start();
-    while (tm.stop() < timems) {
-      for (size_t i = 0; i < batch; i++) {
-        log.start();
-        utils::GemmProblem gp(1, m, n, k);
-        typename Launcher::Param args{gp, {A + i * m * k, k}, {0, 0, &packBs[i]}, {C + i * m * n, n}};
-        parallel::GemmRun<Parallel>(kernel, args, UT_Threading::get());
-        if (log.stop()) {
-          double flops = double(psize) / log.avg_val / 1e6;
-          printf("Threads %d %s %s Flops:%.3f PerCoreFlops:%.3f\n", threads, corestr, log.get_log_str(), flops,
-                 flops / threads);
-        }
-      }
-    }
-  }
-
-  void benchmark_all(size_t m, size_t n, size_t k, size_t batch) {
-    printf("%s %d %d %d %d\n", __FUNCTION__, int(m), int(n), int(k), int(batch));
-    avector<AType> A(m * k * batch);
-    avector<BType> B(k * n * batch);
-    avector<CType> C(m * n * batch), RefC(m * n * batch);
-    fill_buffer_randn(A.data(), m * k, AType(0), AType(255));
-    fill_buffer_randn(B.data(), k * n, BType(-127), BType(127));
-    for (size_t i = 0; i < batch - 1; i++) {
-      memcpy(A.data() + i * m * k, A.data(), m * k * sizeof(AType));
-      memcpy(B.data() + i * n * k, B.data(), n * k * sizeof(BType));
-    }
-    using LOG = timer_statistics_logger<100>;
-    float testtime = 500.f;
-    GetCPUDevice();
-    if (_cd->AMX_INT8()) {
-      request_perm_xtile_data();
-      benchmark<gemm::ICoreRowNAmxint8<32, 32>, LOG>(m, n, k, batch, A.data(), B.data(), C.data(), testtime,
-                                                     _cd->getThreads());
-      benchmark<gemm::ICoreRowNAmxint8<48, 16>, LOG>(m, n, k, batch, A.data(), B.data(), C.data(), testtime,
-                                                     _cd->getThreads());
-      benchmark<gemm::ICoreRowNAmxint8<64, 16>, LOG>(m, n, k, batch, A.data(), B.data(), C.data(), testtime,
-                                                     _cd->getThreads());
-    }
-    if (_cd->AVX512_VNNI()) {
-      benchmark<gemm::ICoreRowNAvx512vnni<48, 8>, LOG>(m, n, k, batch, A.data(), B.data(), C.data(), testtime,
-                                                       _cd->getThreads());
-    }
-    if (_cd->AVX_VNNI()) {
-      if (_cd->isHybrid()) {
-        benchmark<gemm::ICoreRowNAvxvnni<48, 2>, LOG>(m, n, k, batch, A.data(), B.data(), C.data(), testtime,
-                                                      _cd->getThreads());
-        benchmark<gemm::ICoreRowNAvxvnni<48, 2>, LOG>(m, n, k, batch, A.data(), B.data(), C.data(), testtime,
-                                                      _cd->getCores());
-        benchmark<gemm::ICoreRowNAvxvnni<48, 2>, LOG>(m, n, k, batch, A.data(), B.data(), C.data(), testtime,
-                                                      _cd->getPcoreNum());
-        benchmark<gemm::ICoreRowNAvxvnni<24, 4>, LOG>(m, n, k, batch, A.data(), B.data(), C.data(), testtime,
-                                                      _cd->getThreads());
-        benchmark<gemm::ICoreRowNAvxvnni<24, 4>, LOG>(m, n, k, batch, A.data(), B.data(), C.data(), testtime,
-                                                      _cd->getCores());
-        benchmark<gemm::ICoreRowNAvxvnni<24, 4>, LOG>(m, n, k, batch, A.data(), B.data(), C.data(), testtime,
-                                                      _cd->getPcoreNum());
-      } else {
-        benchmark<gemm::ICoreRowNAvxvnni<48, 2>, LOG>(m, n, k, batch, A.data(), B.data(), C.data(), testtime,
-                                                      _cd->getThreads());
-        benchmark<gemm::ICoreRowNAvxvnni<24, 4>, LOG>(m, n, k, batch, A.data(), B.data(), C.data(), testtime,
-                                                      _cd->getThreads());
-      }
-    }
-  }
 };
 #ifdef BTLA_UT_WRAPPER
 static UT_U8S8S32 sUT_U8S8S32;
@@ -306,10 +138,6 @@ class UT_S8S8S32 {
   UT_S8S8S32() {
     UT_START();
     GetCPUDevice();
-#ifdef BTLA_UT_BENCHMARK
-    benchmark_all(1024, 4096, 4096, 32);
-    benchmark_all(2048, 4096, 4096, 32);
-#endif
     if (_cd->AMX_INT8()) {
       request_perm_xtile_data();
       ut<sAMX_INT8_SS>(1, 1, 1);
@@ -359,70 +187,6 @@ class UT_S8S8S32 {
     parallel::GemmRun<Parallel>(launcher, args, UT_Threading::get());
     ut::buffer_error(refC.data(), matC.data(), refC.size(), 0.001f);
   }
-
-  using AType = int8_t;
-  using BType = int8_t;
-  using CType = int;
-  template <typename Core_T, typename LOG_T>
-  void benchmark(int m, int n, int k, int batch, AType* A, BType* B, CType* C, float timems, int threads) {
-    LOG_T log;
-    using Parallel = parallel::gemm::SchedulerBase<Core_T>;
-    using Launcher =
-        wrapper::gemm::LauncherBase<Core_T::ISA, Core_T, prologue_a::gemm::ActivationBase, prologue_b::gemm::WeightPack,
-                                    epilogue::gemm::AccumulatorWriteBackInt32>;
-    Launcher kernel;
-    UT_Threading::set_threads(threads);
-    auto corestr = gemm::CoreAttr::to_str(Core_T::ID);
-    utils::timer<std::chrono::milliseconds> tm;
-    auto tmpB = kernel.mProB.createStorage(n, k);
-    std::vector<storage::gemm::StoragePackedWeight> packBs(batch, 0);
-    std::vector<int8_t> bufB(tmpB.mSize * batch);
-    for (size_t i = 0; i < batch; i++) {
-      packBs[i] = tmpB;
-      packBs[i].assign(bufB.data() + i * tmpB.mSize);
-      kernel.mProB.packWeight(n, k, {B + i * n * k, n, &packBs[i]}, UT_Threading::get());
-    }
-    auto psize = (size_t)m * n * k * 2;
-    tm.start();
-    while (tm.stop() < timems) {
-      for (size_t i = 0; i < batch; i++) {
-        log.start();
-        utils::GemmProblem gp(1, m, n, k);
-        typename Launcher::Param args{gp, {A + i * m * k, k}, {0, 0, &packBs[i]}, {C + i * m * n, n}};
-        parallel::GemmRun<Parallel>(kernel, args, UT_Threading::get());
-        if (log.stop()) {
-          double flops = double(psize) / log.avg_val / 1e6;
-          printf("Threads %d %s %s Flops:%.3f PerCoreFlops:%.3f\n", threads, corestr, log.get_log_str(), flops,
-                 flops / threads);
-        }
-      }
-    }
-  }
-
-  void benchmark_all(size_t m, size_t n, size_t k, size_t batch) {
-    printf("%s %d %d %d %d\n", __FUNCTION__, int(m), int(n), int(k), int(batch));
-    avector<AType> A(m * k * batch);
-    avector<BType> B(k * n * batch);
-    avector<CType> C(m * n * batch), RefC(m * n * batch);
-    fill_buffer_randn(A.data(), m * k, AType(0), AType(255));
-    fill_buffer_randn(B.data(), k * n, BType(-127), BType(127));
-    for (size_t i = 0; i < batch - 1; i++) {
-      memcpy(A.data() + i * m * k, A.data(), m * k * sizeof(AType));
-      memcpy(B.data() + i * n * k, B.data(), n * k * sizeof(AType));
-    }
-    using LOG = timer_statistics_logger<100>;
-    float testtime = 500.f;
-    GetCPUDevice();
-    if (_cd->AMX_INT8()) {
-      request_perm_xtile_data();
-      benchmark<gemm::ICoreRowNAmxint8SS<32, 32>, LOG>(m, n, k, batch, A.data(), B.data(), C.data(), testtime,
-                                                       _cd->getThreads());
-      benchmark<gemm::ICoreRowNAmxint8SS<48, 16>, LOG>(m, n, k, batch, A.data(), B.data(), C.data(), testtime,
-                                                       _cd->getThreads());
-      benchmark<gemm::ICoreRowNAmxint8SS<64, 16>, LOG>(m, n, k, batch, A.data(), B.data(), C.data(), testtime,
-                                                       _cd->getThreads());
-    }
-  }
 };
 #ifdef BTLA_UT_WRAPPER
 static UT_S8S8S32 sUT_S8S8S32;
@@ -433,11 +197,6 @@ class UT_Bf16Bf16Fp32 {
   UT_Bf16Bf16Fp32() {
     UT_START();
     CheckISA(AMX_BF16);
-    request_perm_xtile_data();
-#ifdef BTLA_UT_BENCHMARK
-    benchmark_all(1024, 4096, 4096, 32);
-    benchmark_all(2048, 4096, 4096, 32);
-#endif
     ut<sAMX_BF16>(1, 1, 1);
     ut<sAMX_BF16>(8, 48, 2);
     ut<sAMX_BF16>(8, 4096, 4096);
@@ -468,70 +227,6 @@ class UT_Bf16Bf16Fp32 {
     parallel::GemmRun<Parallel>(launcher, args, UT_Threading::get());
     buffer_error(refC.data(), matC.data(), refC.size(), 0.05f);
   }
-
-  using AType = utils::bf16;
-  using BType = utils::bf16;
-  using CType = float;
-  template <typename Core_T, typename LOG_T>
-  void benchmark(int m, int n, int k, int batch, AType* A, BType* B, CType* C, float timems, int threads) {
-    LOG_T log;
-    using Parallel = parallel::gemm::SchedulerBase<Core_T>;
-    using Launcher =
-        wrapper::gemm::LauncherBase<Core_T::ISA, Core_T, prologue_a::gemm::ActivationBase, prologue_b::gemm::WeightPack,
-                                    epilogue::gemm::AccumulatorWriteBackFp32>;
-    Launcher kernel;
-    UT_Threading::set_threads(threads);
-    auto corestr = gemm::CoreAttr::to_str(Core_T::ID);
-    utils::timer<std::chrono::milliseconds> tm;
-    auto tmpB = kernel.mProB.createStorage(n, k);
-    std::vector<storage::gemm::StoragePackedWeight> packBs(batch, 0);
-    std::vector<int8_t> bufB(tmpB.mSize * batch);
-    for (size_t i = 0; i < batch; i++) {
-      packBs[i] = tmpB;
-      packBs[i].assign(bufB.data() + i * tmpB.mSize);
-      kernel.mProB.packWeight(n, k, {B + i * n * k, n, &packBs[i]}, UT_Threading::get());
-    }
-    auto psize = (size_t)m * n * k * 2;
-    tm.start();
-    while (tm.stop() < timems) {
-      for (size_t i = 0; i < batch; i++) {
-        log.start();
-        utils::GemmProblem gp(1, m, n, k);
-        typename Launcher::Param args{gp, {A + i * m * k, k}, {0, 0, &packBs[i]}, {C + i * m * n, n}};
-        parallel::GemmRun<Parallel>(kernel, args, UT_Threading::get());
-        if (log.stop()) {
-          double flops = double(psize) / log.avg_val / 1e6;
-          printf("Threads %d %s %s Flops:%.3f PerCoreFlops:%.3f\n", threads, corestr, log.get_log_str(), flops,
-                 flops / threads);
-        }
-      }
-    }
-  }
-
-  void benchmark_all(size_t m, size_t n, size_t k, size_t batch) {
-    printf("%s %d %d %d %d\n", __FUNCTION__, int(m), int(n), int(k), int(batch));
-    avector<AType> A(m * k * batch);
-    avector<BType> B(k * n * batch);
-    avector<CType> C(m * n * batch), RefC(m * n * batch);
-    fill_buffer_randn(A.data(), k * m, AType(-0.5f), AType(0.5f));
-    fill_buffer_randn(B.data(), k * n, BType(-0.5f), BType(0.5f));
-    for (size_t i = 0; i < batch - 1; i++) {
-      memcpy(A.data() + i * m * k, A.data(), m * k * sizeof(AType));
-      memcpy(B.data() + i * n * k, B.data(), n * k * sizeof(BType));
-    }
-    using LOG = timer_statistics_logger<100>;
-    float testtime = 500.f;
-    GetCPUDevice();
-    if (_cd->AMX_BF16()) {
-      request_perm_xtile_data();
-      benchmark<gemm::HCoreRowNAmxbf16<32, 32>, LOG>(m, n, k, batch, A.data(), B.data(), C.data(), testtime,
-                                                     _cd->getThreads());
-      benchmark<gemm::HCoreRowNAmxbf16<48, 16>, LOG>(m, n, k, batch, A.data(), B.data(), C.data(), testtime,
-                                                     _cd->getThreads());
-      benchmark<gemm::HCoreRowNAmxbf16<64, 16>, LOG>(m, n, k, batch, A.data(), B.data(), C.data(), testtime,
-                                                     _cd->getThreads());
-    }
-  }
 };
 #ifdef BTLA_UT_WRAPPER
 static UT_Bf16Bf16Fp32 sUT_Bf16Bf16Fp32;
@@ -541,11 +236,6 @@ class UT_Fp16Fp16Fp16 {
  public:
   UT_Fp16Fp16Fp16() {
     UT_START();
-    CheckISA(AVX512_FP16);
-#ifdef BTLA_UT_BENCHMARK
-    benchmark_all(1024, 4096, 4096, 32);
-    benchmark_all(2048, 4096, 4096, 32);
-#endif
     ut<sAVX512_FP16>(1, 1, 1);
     ut<sAVX512_FP16>(8, 48, 2);
     ut<sAVX512_FP16>(8, 4096, 4096);
@@ -574,67 +264,6 @@ class UT_Fp16Fp16Fp16 {
     typename Launcher::Param args{gp, {matAbf16.data(), k}, {matBbf16.data(), n, &packw}, {matC.data(), n}};
     parallel::GemmRun<Parallel>(launcher, args, UT_Threading::get());
     buffer_error(refC.data(), matC.data(), refC.size(), utils::fp16(0.0002f * k));
-  }
-
-  using AType = utils::fp16;
-  using BType = utils::fp16;
-  using CType = utils::fp16;
-  template <typename Core_T, typename LOG_T>
-  void benchmark(int m, int n, int k, int batch, AType* A, BType* B, CType* C, float timems, int threads) {
-    LOG_T log;
-    using Parallel = parallel::gemm::SchedulerBase<Core_T>;
-    using Launcher =
-        wrapper::gemm::LauncherBase<Core_T::ISA, Core_T, prologue_a::gemm::ActivationBase, prologue_b::gemm::WeightPack,
-                                    epilogue::gemm::AccumulatorWriteBackFp16>;
-    Launcher kernel;
-    UT_Threading::set_threads(threads);
-    auto corestr = gemm::CoreAttr::to_str(Core_T::ID);
-    utils::timer<std::chrono::milliseconds> tm;
-    auto tmpB = kernel.mProB.createStorage(n, k);
-    std::vector<storage::gemm::StoragePackedWeight> packBs(batch, 0);
-    std::vector<int8_t> bufB(tmpB.mSize * batch);
-    for (size_t i = 0; i < batch; i++) {
-      packBs[i] = tmpB;
-      packBs[i].assign(bufB.data() + i * tmpB.mSize);
-      kernel.mProB.packWeight(n, k, {B + i * n * k, n, &packBs[i]}, UT_Threading::get());
-    }
-    auto psize = (size_t)m * n * k * 2;
-    tm.start();
-    while (tm.stop() < timems) {
-      for (size_t i = 0; i < batch; i++) {
-        log.start();
-        GemmProblem gp(1, m, n, k);
-        typename Launcher::Param args{gp, {A + i * m * k, k}, {0, 0, &packBs[i]}, {C + i * m * n, n}};
-        parallel::GemmRun<Parallel>(kernel, args, UT_Threading::get());
-        if (log.stop()) {
-          double flops = double(psize) / log.avg_val / 1e6;
-          printf("Threads %d %s %s Flops:%.3f PerCoreFlops:%.3f\n", threads, corestr, log.get_log_str(), flops,
-                 flops / threads);
-        }
-      }
-    }
-  }
-
-  void benchmark_all(size_t m, size_t n, size_t k, size_t batch) {
-    printf("%s %d %d %d %d\n", __FUNCTION__, int(m), int(n), int(k), int(batch));
-    avector<AType> A(m * k * batch);
-    avector<BType> B(k * n * batch);
-    avector<CType> C(m * n * batch), RefC(m * n * batch);
-    fill_buffer_randn(A.data(), k * m, AType(-0.5f), AType(0.5f));
-    fill_buffer_randn(B.data(), k * n, AType(-0.5f), AType(0.5f));
-    for (size_t i = 0; i < batch - 1; i++) {
-      memcpy(A.data() + i * m * k, A.data(), m * k * sizeof(AType));
-      memcpy(B.data() + i * n * k, B.data(), n * k * sizeof(BType));
-    }
-    using LOG = timer_statistics_logger<100>;
-    float testtime = 500.f;
-    GetCPUDevice();
-    if (_cd->AVX512_FP16()) {
-      benchmark<sAVX512_FP16, LOG>(m, n, k, batch, A.data(), B.data(), C.data(), testtime, _cd->getThreads());
-      benchmark<gemm::HCoreRowNAvx512fp16<64, 12>, LOG>(m, n, k, batch, A.data(), B.data(), C.data(), testtime,
-                                                        _cd->getThreads());
-      benchmark<sAVX512_FP16, LOG>(m, n, k, batch, A.data(), B.data(), C.data(), testtime, _cd->getThreads());
-    }
   }
 };
 #ifdef BTLA_UT_WRAPPER
